Refactor mlas unittest.
